[CAS-582] Deleting Old ChannelControllerImplEventTest
